LayoutTests/inspector/canvas/setShaderProgramHighlighted-expected.txt

Test highlighting of shader programs.


== Running test suite: Canvas.setShaderProgramHighlighted
-- Running test case: Canvas.setShaderProgramHighlighted.highlightedDrawArrays
PASS: Highlighting a shader program should change the content.

-- Running test case: Canvas.setShaderProgramHighlighted.highlightedDrawElements
PASS: Highlighting a shader program should change the content.

-- Running test case: Canvas.setShaderProgramHighlighted.unhighlightedDrawArrays
PASS: Unhighlighting a shader program should change the content.

-- Running test case: Canvas.setShaderProgramHighlighted.unhighlightedDrawElements
PASS: Unhighlighting a shader program should change the content.

-- Running test case: Canvas.setShaderProgramHighlighted.invalidProgramId
PASS: Should produce an error.
Error: No shader program for given identifier.
